Separate ConsolePanel and ConsoleWidget. Add stub for ConsolePager.
