+++ src/console/panel.ts
@@ -1,0 +1,164 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+import {
+  ISession
+} from 'jupyter-js-services';
+
+import {
+  Message
+} from 'phosphor-messaging';
+
+import {
+  Widget
+} from 'phosphor-widget';
+
+import {
+  SplitPanel
+} from 'phosphor-splitpanel';
+
+import {
+  showDialog
+} from '../dialog';
+
+import {
+  RenderMime
+} from '../rendermime';
+
+import {
+  ConsoleWidget
+} from './widget';
+
+
+/**
+ * The class name added to console panels.
+ */
+const CONSOLE_PANEL = 'jp-Console-panel';
+
+
+/**
+ * A panel which contains a toolbar and a console.
+ */
+export
+class ConsolePanel extends SplitPanel {
+  /**
+   * Construct a console panel.
+   */
+  constructor(options: ConsolePanel.IOptions) {
+    super();
+    this.addClass(CONSOLE_PANEL);
+    this._console = new ConsoleWidget({
+      session: options.session,
+      rendermime: options.rendermime
+    });
+    this.addChild(this._console);
+  }
+
+  /**
+   * The console widget used by the panel.
+   *
+   * #### Notes
+   * This is a read-only property.
+   */
+  get content(): ConsoleWidget {
+    return this._console;
+  }
+
+  /**
+   * Dispose of the resources held by the widget.
+   */
+  dispose(): void {
+    if (this.isDisposed) {
+      return;
+    }
+    this._console.dispose();
+    this._console = null;
+    super.dispose();
+  }
+
+  /**
+   * Handle the DOM events for the widget.
+   *
+   * @param event - The DOM event sent to the widget.
+   *
+   * #### Notes
+   * This method implements the DOM `EventListener` interface and is
+   * called in response to events on the dock panel's node. It should
+   * not be called directly by user code.
+   */
+  handleEvent(event: Event): void {
+    switch (event.type) {
+    case 'click':
+      let prompt = this.content.prompt;
+      if (prompt) {
+        prompt.focus();
+      }
+      break;
+    default:
+      break;
+    }
+  }
+
+  /**
+   * Handle `after_attach` messages for the widget.
+   */
+  protected onAfterAttach(msg: Message): void {
+    this.content.node.addEventListener('click', this);
+  }
+
+  /**
+   * Handle `before_detach` messages for the widget.
+   */
+  protected onBeforeDetach(msg: Message): void {
+    this.content.node.removeEventListener('click', this);
+  }
+
+  /**
+   * Handle `'close-request'` messages.
+   */
+  protected onCloseRequest(msg: Message): void {
+    let session = this.content.session;
+    if (!session.kernel) {
+      this.dispose();
+    }
+    session.kernel.getKernelSpec().then(spec => {
+      let name = spec.display_name;
+      return showDialog({
+        title: 'Shut down kernel?',
+        body: `Shut down ${name}?`
+      });
+    }).then(value => {
+      if (value && value.text === 'OK') {
+        return session.shutdown();
+      }
+    }).then(() => {
+      super.onCloseRequest(msg);
+      this.dispose();
+    });
+  }
+
+  private _console: ConsoleWidget = null;
+}
+
+
+/**
+ * A namespace for ConsolePanel statics.
+ */
+export
+namespace ConsolePanel {
+  /**
+   * The initialization options for a console panel.
+   */
+  export
+    interface IOptions {
+    /**
+     * The session for the console panel.
+     */
+    session: ISession;
+
+    /**
+     * The mime renderer for the console panel.
+     */
+    rendermime: RenderMime<Widget>;
+  }
+}
