+++ src/console/widget.ts
@@ -2,7 +2,7 @@
 // Distributed under the terms of the Modified BSD License.
 
 import {
-  IKernel, ISession, KernelMessage
+  ISession, KernelMessage
 } from 'jupyter-js-services';
 
 import {
@@ -26,8 +26,12 @@ import {
 } from 'phosphor-widget';
 
 import {
-  PanelLayout, Panel
+  PanelLayout
 } from 'phosphor-panel';
+
+import {
+  SplitPanel
+} from 'phosphor-splitpanel';
 
 import {
   CodeCellWidget, CodeCellModel, RawCellModel, RawCellWidget
@@ -64,11 +68,6 @@ import {
 const CONSOLE_CLASS = 'jp-Console';
 
 /**
- * The class name added to console panels.
- */
-const CONSOLE_PANEL = 'jp-Console-panel';
-
-/**
  * The class name added to the console banner.
  */
 const BANNER_CLASS = 'jp-Console-banner';
@@ -77,134 +76,6 @@ const BANNER_CLASS = 'jp-Console-banner';
  * The class name of the active prompt
  */
 const PROMPT_CLASS = 'jp-Console-prompt';
-
-
-/**
- * A panel which contains a toolbar and a console.
- */
-export
-class ConsolePanel extends Panel {
-  /**
-   * Construct a console panel.
-   */
-  constructor(options: ConsolePanel.IOptions) {
-    super();
-    this.addClass(CONSOLE_PANEL);
-    this._console = new ConsoleWidget({
-      session: options.session,
-      rendermime: options.rendermime
-    });
-    this.addChild(this._console);
-  }
-
-  /**
-   * The console widget used by the panel.
-   *
-   * #### Notes
-   * This is a read-only property.
-   */
-  get content(): ConsoleWidget {
-    return this._console;
-  }
-
-  /**
-   * Dispose of the resources held by the widget.
-   */
-  dispose(): void {
-    if (this.isDisposed) {
-      return;
-    }
-    this._console.dispose();
-    this._console = null;
-    super.dispose();
-  }
-
-  /**
-   * Handle the DOM events for the widget.
-   *
-   * @param event - The DOM event sent to the widget.
-   *
-   * #### Notes
-   * This method implements the DOM `EventListener` interface and is
-   * called in response to events on the dock panel's node. It should
-   * not be called directly by user code.
-   */
-  handleEvent(event: Event): void {
-    switch (event.type) {
-    case 'click':
-      let prompt = this.content.prompt;
-      if (prompt) {
-        prompt.focus();
-      }
-      break;
-    default:
-      break;
-    }
-  }
-
-  /**
-   * Handle `after_attach` messages for the widget.
-   */
-  protected onAfterAttach(msg: Message): void {
-    this.content.node.addEventListener('click', this);
-  }
-
-  /**
-   * Handle `before_detach` messages for the widget.
-   */
-  protected onBeforeDetach(msg: Message): void {
-    this.content.node.removeEventListener('click', this);
-  }
-
-  /**
-   * Handle `'close-request'` messages.
-   */
-  protected onCloseRequest(msg: Message): void {
-    let session = this.content.session;
-    if (!session.kernel) {
-      this.dispose();
-    }
-    session.kernel.getKernelSpec().then(spec => {
-      let name = spec.display_name;
-      return showDialog({
-        title: 'Shut down kernel?',
-        body: `Shut down ${name}?`
-      });
-    }).then(value => {
-      if (value && value.text === 'OK') {
-        return session.shutdown();
-      }
-    }).then(() => {
-      super.onCloseRequest(msg);
-      this.dispose();
-    });
-  }
-
-  private _console: ConsoleWidget = null;
-}
-
-
-/**
- * A namespace for ConsolePanel statics.
- */
-export
-namespace ConsolePanel {
-  /**
-   * The initialization options for a console panel.
-   */
-  export
-    interface IOptions {
-    /**
-     * The session for the console panel.
-     */
-    session: ISession;
-
-    /**
-     * The mime renderer for the console panel.
-     */
-    rendermime: RenderMime<Widget>;
-  }
-}
 
 
 /**
